move Game into Starter
