+++ game/test/controller/LibgdxSetupTest.java
@@ -8,6 +8,7 @@ import org.mockito.Mockito;
 import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
+import starter.Game;
 import starter.LibgdxSetup;
 
 @RunWith(PowerMockRunner.class)
