+++ game/test/starter/GameTest.java
@@ -1,4 +1,4 @@
-package controller;
+package starter;
 
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.when;
